+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v1.0.0-rc9
 ### Changes
 * Updated auralog and entgo dependencies to the latest version
+* Switched to a faster json package (go-json) instead of using encoding/json
 
 ## v1.0.0-rc8
 ### Added